+++ app/src/main/java/elite/intel/ai/brain/actions/handlers/commands/OpenFleetCarrierManagementHandler.java
@@ -4,6 +4,7 @@ import com.google.gson.JsonObject;
 import elite.intel.ai.hands.events.GameInputEvent;
 import elite.intel.gameapi.GameControllerBus;
 import elite.intel.session.Status;
+import elite.intel.util.SleepNoThrow;
 
 import static elite.intel.ai.hands.Bindings.GameCommand.*;
 
@@ -13,20 +14,30 @@ public class OpenFleetCarrierManagementHandler implements CommandHandler {
     private final Status status = Status.getInstance();
 
     /// not a sure fire. assumes default UI selection. will fail often.
-    @Override public void handle(String action, JsonObject params, String responseText) {
+    @Override
+    public void handle(String action, JsonObject params, String responseText) {
         UiNavCommon.close();
-        if(status.isOnFoot()){
+        if (status.isOnFoot()) {
             GameControllerBus.publish(new GameInputEvent(BINDING_ON_FOOT_WHEEL.getGameBinding(), 500));
             GameControllerBus.publish(new GameInputEvent(BINDING_UI_RIGHT.getGameBinding(), 0));
             GameControllerBus.publish(new GameInputEvent(BINDING_UI_DOWN.getGameBinding(), 0));
             GameControllerBus.publish(new GameInputEvent(BINDING_UI_DOWN.getGameBinding(), 0));
             GameControllerBus.publish(new GameInputEvent(BINDING_UI_DOWN.getGameBinding(), 0));
             GameControllerBus.publish(new GameInputEvent(BINDING_ACTIVATE.getGameBinding(), 0));
-        } else if(status.isInMainShip()){
+        } else if (status.isInMainShip()) {
             GameControllerBus.publish(new GameInputEvent(BINDING_FOCUS_INTERNAL_PANEL.getGameBinding(), 0));
+            GameControllerBus.publish(new GameInputEvent(BINDING_UI_UP.getGameBinding(), 0));
+            SleepNoThrow.sleep(100);
+            GameControllerBus.publish(new GameInputEvent(BINDING_UI_LEFT.getGameBinding(), 0));
+            SleepNoThrow.sleep(100);
+            GameControllerBus.publish(new GameInputEvent(BINDING_UI_UP.getGameBinding(), 0));
+            SleepNoThrow.sleep(100);
+            GameControllerBus.publish(new GameInputEvent(BINDING_UI_UP.getGameBinding(), 0));
+            SleepNoThrow.sleep(100);
             GameControllerBus.publish(new GameInputEvent(BINDING_UI_DOWN.getGameBinding(), 0));
+            SleepNoThrow.sleep(100);
             GameControllerBus.publish(new GameInputEvent(BINDING_ACTIVATE.getGameBinding(), 0));
-        } else if(status.isInSrv()){
+        } else if (status.isInSrv()) {
             GameControllerBus.publish(new GameInputEvent(BINDING_FOCUS_INTERNAL_PANEL_BUGGY.getGameBinding(), 0));
             GameControllerBus.publish(new GameInputEvent(BINDING_UI_DOWN.getGameBinding(), 0));
             GameControllerBus.publish(new GameInputEvent(BINDING_ACTIVATE.getGameBinding(), 0));