+++ app/src/main/java/elite/intel/gameapi/journal/subscribers/StartJumpSubscriber.java
@@ -9,8 +9,6 @@ import elite.intel.search.edsm.dto.DeathsDto;
 import elite.intel.search.edsm.dto.TrafficDto;
 import elite.intel.session.PlayerSession;
 
-import static elite.intel.util.StringUtls.isFuelStarClause;
-
 @SuppressWarnings("unused")
 public class StartJumpSubscriber {
 
@@ -21,13 +19,12 @@ public class StartJumpSubscriber {
             TrafficDto trafficDto = EdsmApiClient.searchTraffic(event.getStarSystem());
             DeathsDto deathsDto = EdsmApiClient.searchDeaths(event.getStarSystem());
             StringBuilder sb = new StringBuilder();
-            sb.append("We are traveling through hyperspace on route to ");
+                sb.append("In route to ");
             sb.append(event.getStarSystem());
             sb.append(", ");
-            sb.append("Star Class: ");
-            sb.append(event.getStarClass());
-            sb.append(", ");
-            sb.append(isFuelStarClause(event.getStarClass()));
+                sb.append("Star class: ");
+                sb.append(event.getStarClass());
+                sb.append(". ");
                 if (trafficDto.getData() != null && trafficDto.getData().getTraffic() != null && trafficDto.getData().getTraffic().getTotal() > 0) {
                 sb.append(" Traffic data: " + trafficDto.getData().getTraffic().toYaml());
             }
@@ -39,11 +36,9 @@ public class StartJumpSubscriber {
             playerSession.clearGenusPaymentAnnounced();
             if (playerSession.isRouteAnnouncementOn()) {
                 String instructions = """
-                Notify User about the star system we are traveling to.
-                    - IMPORTANT: Mention star class and if the star is scoopable for hydrogen fuel or not.
-                Example 1: In route to X star class Y, scoopable for fuel.
-                Example 2: In route to X star class Y, WARNING! No fuel available.
-                
+                        Notify User about the star system we are traveling to using this exact format.
+                        Example: In route to <name>, <class> class star.
+                        
                 Data may include traffic and fatalities.
                 Traffic total,weekly and daily indicates number of ships traveled through this system. Deaths data indicates number of ships lost in this system.
                 Example: Traffic: total X, weekly Y, daily Z. Deaths: total A, weekly B, daily C.